SAF USB parser update, Error cases
:Release Notes:
SAF USB parser update, Error cases

:Detailed Notes:
SAF USB parser update, Error cases

:Testing Performed:
Build succesful and verified

:QA Notes:
N/A

:Issues Addressed:
[PLAT-124860] [OSE SAF USB] Modify USB parser

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.storageaccess/+/304345
Tested-by: RajeshGopu IV <rajeshgopu.iv@lge.com>

+++ src/providers/gdrive/GDriveProvider.cpp
@@ -225,7 +225,7 @@ void GDriveProvider::list(std::shared_ptr<RequestData> reqData)
             if (-1 == file.get_fileSize())
                 contentObj.put("size", 0);
             else
-                contentObj.put("size", (totalspace - file.get_fileSize()));
+                contentObj.put("size", (int)(totalspace - file.get_fileSize()));
             contentsObj.append(contentObj);
         }
         respObj.put("returnValue", true);
@@ -278,8 +278,8 @@ void GDriveProvider::getProperties(std::shared_ptr<RequestData> reqData)
     respObj.put("deletable", true);
     if (path == "root")
     {
-        respObj.put("totalSpace", totalspace / 1000000);
-        respObj.put("freeSpace", freespace / 1000000);
+        respObj.put("totalSpace", (int)(totalspace / 1000000));
+        respObj.put("freeSpace", (int)(freespace / 1000000));
     }
     reqData->cb(respObj, reqData->subs);
 }
